README.md: Improve the MSYS2 instructions I originally committed, and improve the instructions below.
It's better to let MSYS2 use its default CMake generator, which
is Ninja.

The instructions below that assumed that Makefiles were being used,
because they said "make install", but we don't want to use Make on
MSYS2, so I fixed that.

I also noticed that the instructions below could be improved in other
ways.  There was typo (picootol), some misplaced commas,
shorthand (ELF->UF2), and passive voice.
It did not mention the option of setting picotool_DIR as an
environment variable (which is very handy), and it was setting
picotool_DIR equal to the path to the executable instead of the
directory, which seems wrong.