Fix compilation on FreeBSD 13.2 and later. (#133)
The main change is to portable_endian.h which seems to be out-of-date
and not maintained upstream.  This change also impacts OpenBSD,
but a check of current OpenBSD git repository suggests it is correct there
also.
Finally, <cuchar> is excluded in main.cpp (as it already was for __Apple__);
this will probably not be needed in later FreeBSD releases once <cuchar>
has been picked up from more recent llvm.