fix the order of the udev rules
The current udev rule name will not actually apply the uaccess tag
in time for it to have the intended effect of setting the
permissions to allow the local seated user to access the device.
To do this, the rule must have a prefix <73 due to
"73-seat-late.rules" being the rule that actually uses the tag to
set the appropriate permissions. This change moves the rule to a
position where it will function as originally intended.